+++ src/expr/src/scalar/func.rs
@@ -3332,6 +3332,7 @@ pub enum UnaryFunc {
         // The expression to cast List1's elements to List2's elements' type
         cast_expr: Box<MirScalarExpr>,
     },
+    CastArrayToListOneDim(CastArrayToListOneDim),
     CastMapToString {
         ty: ScalarType,
     },
@@ -3534,6 +3535,7 @@ derive_unary!(
     CastStringToVarChar,
     CastCharToString,
     CastUuidToString,
+    CastArrayToListOneDim,
     Cos,
     Acos,
     Cosh,
@@ -3722,6 +3724,7 @@ impl UnaryFunc {
             | CastIntervalToTime(_)
             | NegInterval(_)
             | CastUuidToString(_)
+            | CastArrayToListOneDim(_)
             | CastTimestampToString(_)
             | CastTimestampTzToString(_)
             | CastTimestampToDate(_)
@@ -3946,6 +3949,7 @@ impl UnaryFunc {
             | CastIntervalToTime(_)
             | NegInterval(_)
             | CastUuidToString(_)
+            | CastArrayToListOneDim(_)
             | CastTimestampToString(_)
             | CastTimestampTzToString(_)
             | CastTimestampToDate(_)
@@ -4195,6 +4199,7 @@ impl UnaryFunc {
             | CastIntervalToTime(_)
             | NegInterval(_)
             | CastUuidToString(_)
+            | CastArrayToListOneDim(_)
             | CastTimestampToString(_)
             | CastTimestampTzToString(_)
             | CastTimestampToDate(_)
@@ -4465,6 +4470,7 @@ impl UnaryFunc {
             | CastIntervalToTime(_)
             | NegInterval(_)
             | CastUuidToString(_)
+            | CastArrayToListOneDim(_)
             | CastTimestampToString(_)
             | CastTimestampTzToString(_)
             | CastTimestampToDate(_)
@@ -5953,6 +5959,7 @@ mod test {
         let mut rti = lowertest::ReflectedTypeInfo::default();
         UnaryFunc::add_to_reflected_type_info(&mut rti);
         for (variant, (_, f_types)) in rti.enum_dict["UnaryFunc"].iter() {
+            println!("f_types {:?}", f_types);
             if f_types.is_empty() {
                 let unary_unit_variant: UnaryFunc =
                     serde_json::from_str(&format!("\"{}\"", variant)).unwrap();